[refactor]: 메인 페이지 util 함수 분리
